[LibGit2] isolate tests from ~/.gitconfig
Settings such as insteadOf can break the tests.
