+++ stdlib/LibGit2/test/runtests.jl
@@ -1,4 +1,6 @@
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
-include("libgit2.jl")
-include("online.jl")
+using Test
+@testset verbose=true "LibGit2 $test" for test in eachline(joinpath(@__DIR__, "testgroups"))
+    include("$test.jl")
+end
